Rename module data_clerk -> clerk
